board: microchip: sev-kit: read device tree overlays from the design
If available, read the device tree overlays from the design
structure and create the env variables for number of overlays,
device tree overlays image address and total size of device
tree overlays. Later these env variables are used to apply
overlays to the device tree.

Support for this is already present on Icicle, so copy across the
implementation from Icicle.

Signed-off-by: Shravan Chippa <shravan.chippa@microchip.com>
[conor: re-wrote commit message slightly]
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>